+++ Gemfile
@@ -12,3 +12,7 @@ gem "soda-ruby"
 group :jekyll_plugins do
   gem "jekyll-contentful-data-import"
 end
+
+group :development do
+  gem "foreman"
+end
