+++ _config/pec.yml
@@ -1,0 +1,13 @@
+id: pec
+title: Palmer Events Center
+url: "http://beta.palmereventscenter.com"
+
+calendar:
+  location: "Palmer Events Center" # Socrata event data location value
+
+google_analytics_id: TBD
+
+# Build settings
+destination: _site/pec
+assets:
+  cache: .asset-cache/pec
