+++ _layouts/floor_plan.html
@@ -5,43 +5,3 @@ layout: default
 <div class="usa-grid">
   <h1>Floor Plans</h1>
 </div>
-
-<div class="usa-grid acc-floor-plan">
-  <aside class="usa-width-one-fourth">
-
-    <ul class="usa-sidenav-list">
-      <li><a href="/facilities/plans/"{% if page.url == "/facilities/plans/" %} class="usa-current"{% endif %}>Center Overview</a></li>
-
-      {% assign levels = site.plans | where_exp: "page", "page.slug contains 'level'" | sort: "slug" %}
-
-      {% for level in levels %}
-        <li><a href="{{ level.url }}"{% if page.url == level.url %} class="usa-current"{% endif %}>{{ level.title }}</a></li>
-      {% endfor %}
-
-      {% for doc in site.plans %}
-        {% unless doc.slug contains 'level' %}
-          <li>
-            <a href="{{ doc.url }}"{% if page.url == doc.url %} class="usa-current"{% endif %}>{{ doc.title }}</a>
-
-            {% if doc.docs %}
-              <ul class="usa-sidenav-sub_list">
-                {% assign child_docs = doc.docs | sort: "title" %}
-                {% for child_doc in child_docs %}
-                  <a href="{{ child_doc.url }}"{% if page.url == child_doc.url %} class="usa-current"{% endif %}>{{ child_doc.title }}</a>
-                {% endfor %}
-              </ul>
-            {% endif %}
-
-          </li>
-        {% endunless %}
-      {% endfor %}
-
-    </ul>
-
-  </aside>
-
-  <div class="usa-width-three-fourths">
-    <h3>{{ page.contentful.title }}</h3>
-    {% include components/content_blocks.html blocks=page.contentful.contentBlocks %}
-  </div>
-</div>
