--- _plugins/mappers/content_block_mapper.rb
@@ -1,9 +1,0 @@
-class ContentBlockMapper < Jekyll::Contentful::Mappers::Base
-
-  def map
-    result = super
-    result["type"] = self.entry.content_type.id
-    result
-  end
-
-end
