+++ lib/mappers/content_block_mapper.rb
@@ -1,0 +1,11 @@
+module Mappers
+  class ContentBlockMapper < Jekyll::Contentful::Mappers::Base
+
+    def map
+      result = super
+      result["type"] = self.entry.content_type.id
+      result
+    end
+
+  end
+end
