+++ s3_website.yml
@@ -1,6 +1,6 @@
 s3_id: <%= ENV["AWS_ACCESS_KEY_ID"] %>
 s3_secret: <%= ENV["AWS_SECRET_ACCESS_KEY"] %>
-s3_bucket: alpha.austinconventioncenter.com
+s3_bucket: <%= ENV["S3_BUCKET"] %>
 
 # Below are examples of all the available configurations.
 # See README for more detailed info on each of them.
